# backend/model/src/main/java/eu/viandeendirect/model/OrderStatus.java
package eu.viandeendirect.model;

import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonValue;
import jakarta.annotation.Generated;

/**
 * Gets or Sets OrderStatus
 */

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
public enum OrderStatus {
  
  ITEMS_SELECTED("ITEMS_SELECTED"),

  PAYMENT_COMPLETED("PAYMENT_COMPLETED"),

  PAYEMENT_VALIDATED("PAYEMENT_VALIDATED"),
  
  PAYMENT_FAILED("PAYMENT_FAILED"),
  
  DELIVERED("DELIVERED");

  private String value;

  OrderStatus(String value) {
    this.value = value;
  }

  @JsonValue
  public String getValue() {
    return value;
  }

  @Override
  public String toString() {
    return String.valueOf(value);
  }

  @JsonCreator
  public static OrderStatus fromValue(String value) {
    for (OrderStatus b : OrderStatus.values()) {
      if (b.value.equals(value)) {
        return b;
      }
    }
    throw new IllegalArgumentException("Unexpected value '" + value + "'");
  }
}