# openapi/openapi.yml
openapi: 3.0.2
info:
  title: viandeendirect.eu
  version: 0.0.1
  description: This API allows selling and buying food through the shortest path between
    farmers and customers.
servers:
  - url: http://localhost:8080
security:
  - oAuth2ForViandeEnDirect:
    - read
    - write
paths:
...
  /customers:
    summary: Path used to manage customers
    get:
      parameters:
        - in: query
          name: email
          schema: 
            type: string
          description: "the email of the Customer to load"
      security:
        - oAuth2ForViandeEnDirect: [read]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
          description: Successful response - returns a `Customer` entities.
      operationId: getCustomer
      summary: Get a customer identified by its email.
      description: Gets a `Customer` entities.
    post:
      security:
        - oAuth2ForViandeEnDirect: [write]
      requestBody:
        description: A new `customer` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
          description: Successful response - returns the `Customer` entity created.
      operationId: createCustomer
      summary: Create a new customer
...
  /orders:
    summary: Path used to manage orders.
    description: "The REST endpoint/path used to list and create zero or more `Order`\
      \ entities.  This path contains a `GET` and `POST` operation to perform the\
      \ list and create tasks, respectively."
    post:
      security:
        - oAuth2ForViandeEnDirect: [write]
      requestBody:
        description: A new `Order` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: Successful response.
      operationId: createOrder
      summary: Create a Order
      description: Creates a new instance of a `Order`.
  /orders/{orderId}:
    summary: Path used to manage a single Order.
    get:
      security:
        - oAuth2ForViandeEnDirect: [read]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: Successful response - returns a single `Order`.
      operationId: getOrder
      summary: Get a Order
      description: Gets the details of a single instance of a `Order`.
    parameters:
    - name: orderId
      description: A unique identifier for a `Order`.
      schema:
        type: integer
      in: path
      required: true
  /orders/payment:
    summary: Path used to manage an order payment.
    post:
      security:
        - oAuth2ForViandeEnDirect: [write]
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: Successful response - returns a single `Order` with its `StripePayment`.
      operationId: createOrderPayment
      summary: Creates a payment for the order given as a payload
      description: Creates an order and its payment that should be processes afterward on payment platform.
      requestBody:
        description: A new `Order` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
components:
  securitySchemes:
    oAuth2ForViandeEnDirect:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8180/realms/viandeendirect/protocol/openid-connect/auth
          tokenUrl: http://localhost:8180/realms/viandeendirect/protocol/openid-connect/token
          scopes:
            read: Grants read access
            write: Grants write access
            admin: Grants access to admin operations
  schemas:
    PackageLot:
      description: "A lot of packages"
      type: object
      properties:
        id:
          type: integer
          x-field-extra-annotation: '@jakarta.persistence.Id'
        production:
          $ref: '#/components/schemas/Production'
          description: "The production from which the packages are prepared"
          x-field-extra-annotation: '@jakarta.persistence.ManyToOne'
        label:
          description: "Short label of the package template"
          type: string
        description:
          description: "Long description of the package template"
          type: string
        photo:
          description: "The photo of the package template, encoded as base64"
          type: string
        netWeight:
          description: ""
          type: number
          format: float
        unitPrice:
          description: |-
            Price per unit (weight unit).
            Must be multiplied by the weight to get the total price for one package.
            Includes taxes.
          type: number
          format: float
        quantity:
          description: "the quantity of packages for this batch initially on sale"
          type: integer
        quantitySold:
          description: "the quantity of packages for this batch already sold"
          type: integer
      x-class-extra-annotation: '@jakarta.persistence.Entity @jakarta.persistence.Table(name
        = "packages")'
...
    User:
      description: ""
      type: object
      properties:
        id:
          description: ""
          type: integer
          x-field-extra-annotation: '@jakarta.persistence.Id'
        email:
          type: string
          format: email
        phone:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      x-class-extra-annotation: '@jakarta.persistence.Entity @jakarta.persistence.Table(name
        = "users")'
    Customer:
      description: ""
      type: object
      x-class-extra-annotation: '@jakarta.persistence.Entity @jakarta.persistence.Table(name = "customers")'
      properties:
        user:
          $ref: '#/components/schemas/User'
          description: ""
          x-field-extra-annotation: '@jakarta.persistence.OneToOne'
        id:
          description: ""
          type: integer
          x-field-extra-annotation: '@jakarta.persistence.Id'
    Invoice:
      description: ""
      type: object
      properties:
        id:
          description: ""
          type: integer
          x-field-extra-annotation: '@jakarta.persistence.Id'
      x-class-extra-annotation: '@jakarta.persistence.Entity @jakarta.persistence.Table(name = "invoices")'
    Production:
      description: ""
      type: object
      properties:
        id:
          description: ""
          type: integer
          x-field-extra-annotation: '@jakarta.persistence.Id'
        productionType:
          type: string
          enum: ['BeefProduction', 'HonneyProduction']
        sales:
          description: ""
          type: array
          items:
            $ref: '#/components/schemas/Sale'
          x-field-extra-annotation: '@jakarta.persistence.ManyToMany(mappedBy = "productions")'
        producer:
          $ref: '#/components/schemas/Producer'
          description: ""
          x-field-extra-annotation: '@jakarta.persistence.ManyToOne'
        lots:
          description: ""
          type: array
          items:
            $ref: '#/components/schemas/PackageLot'
          x-field-extra-annotation: '@jakarta.persistence.OneToMany(mappedBy = "production")'
      discriminator:
        propertyName: 'productionType'
        mapping:
          'BeefProduction': '#/components/schemas/BeefProduction'
      x-class-extra-annotation: '@jakarta.persistence.Entity @jakarta.persistence.Inheritance(strategy = jakarta.persistence.InheritanceType.TABLE_PER_CLASS)'
    BeefProduction:
      allOf:
        - $ref: '#/components/schemas/Production'
      description: "a production a beef meet, obtained by slaughtering and cutting an entire animal"
      type: object
      properties:
        animalIdentifier:
          description: "official identifier of the animal"
          type: string
        animalType:
          description: "type of animal"
          type: string
          enum: ['BEEF_COW', 'BEEF_HEIFER', 'BEEF_BULL', 'BEEF_VEAL']
        cattleBreed:
          description: "breed of the animal"
          type: string
          enum: ['LIMOUSINE', 'CHAROLAISE']
        birthDate:
          description: "date when the animal is born"
          type: string
          format: date
        birthPlace:
          description: "place where the animal is born"
          type: string
        birthFarm:
          description: "farming company where the animal is born"
          type: string
        warmCarcassWeight:
          description: "weight of the animal just after being killed, in kilograms"
          type: integer
        meatWeight:
          description: "weight of meat predicted, in kilograms"
          type: integer
        slaughterDate:
          description: "date when the animal has been killed"
          type: string
          format: date
        slaughterPlace:
          description: "the place where the animal has been killed"
          type: string
        slaughterHouse:
          description: "the company where the animal has been killed"
          type: string
        cuttingDate:
          description: "date when the animal has been cutted"
          type: string
          format: date
        cuttingPlace:
          description: "the place where the animal has been cutted"
          type: string
        cuttingButcher:
          description: "butcher company where the animal has been cutted"
          type: string
        labelRougeCertified:
          description: "true if the animal breeding process matches the french 'label rouge certification"
          type: boolean
      x-class-extra-annotation: '@jakarta.persistence.Entity @jakarta.persistence.Table(name = "beef_productions")'
...
    OrderItem:
      description: ""
      required:
      - package
      type: object
      properties:
        id:
          description: ""
          type: integer
          x-field-extra-annotation: '@jakarta.persistence.Id'
        order:
          $ref: '#/components/schemas/Order'
          x-field-extra-annotation: '@jakarta.persistence.ManyToOne'
        packageLot:
          $ref: '#/components/schemas/PackageLot'
          x-field-extra-annotation: '@jakarta.persistence.OneToOne'
        unitPrice:
          description: |-
            Price per unit ordered.
            Must be multiplied per quantity to get the total price.
            Includes taxes.
          type: number
          format: float
        quantity:
          description: ""
          type: integer
      x-class-extra-annotation: '@jakarta.persistence.Entity @jakarta.persistence.Table(name = "orderItems")'
    Order:
      description: ""
      required:
      - customer
      type: object
      properties:
        id:
          description: ""
          type: integer
          x-field-extra-annotation: '@jakarta.persistence.Id'
        invoice:
          $ref: '#/components/schemas/Invoice'
          description: ""
          x-field-extra-annotation: '@jakarta.persistence.OneToOne'
        items:
          description: ""
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          x-field-extra-annotation: '@jakarta.persistence.OneToMany(mappedBy = "order")'
        customer:
          $ref: '#/components/schemas/Customer'
          x-field-extra-annotation: '@jakarta.persistence.OneToOne'
        sale:
          $ref: '#/components/schemas/Sale'
          x-field-extra-annotation: '@jakarta.persistence.ManyToOne'
        payment:
          $ref: '#/components/schemas/StripePayment'
          x-field-extra-annotation: '@jakarta.persistence.OneToOne'
        status:
          $ref: '#/components/schemas/OrderStatus'
      x-class-extra-annotation: '@jakarta.persistence.Entity @jakarta.persistence.Table(name = "orders")'
    Sale:
      description: ""
      type: object
      properties:
        id:
          type: integer
          x-field-extra-annotation: '@jakarta.persistence.Id'
        seller:
          $ref: '#/components/schemas/Producer'
          x-field-extra-annotation: '@jakarta.persistence.ManyToOne'
        productions:
          description: ""
          type: array
          items:
            $ref: '#/components/schemas/Production'
            x-field-extra-annotation: '@jakarta.persistence.ManyToMany'
        orders:
          description: ""
          type: array
          items:
            $ref: '#/components/schemas/Order'
            x-field-extra-annotation: '@jakarta.persistence.OneToMany(mappedBy = "sale")'
        deliveryStart:
          description: ""
          type: string
          format: date-time
        deliveryStop:
          description: ""
          type: string
          format: date-time
        deliveryAddressName:
          type: string
        deliveryAddressLine1:
          type: string
        deliveryAddressLine2:
          type: string
        deliveryCity:
          type: string
        deliveryZipCode:
          type: string
      x-class-extra-annotation: '@jakarta.persistence.Entity @jakarta.persistence.Table(name = "sales")'
    StripeAccount:
      description: an acount for Stripe payment service
      type: object
      properties:
        id:
          type: integer
          x-field-extra-annotation: '@jakarta.persistence.Id'
          description: the Stripe account id given by viandeendirect.eu
        stripeId:
          type: string
          description: the Stripe account id given by Stripe service
        detailsSubmitted:
          x-field-extra-annotation: '@jakarta.persistence.Transient'
          description: true if additionnal informations of the account owner have been fulfilles in Stripe platform
          type: boolean
        accountLink:
          x-field-extra-annotation: '@jakarta.persistence.Transient'
          description: the link to the account on Stripe platform
          type: string
    StripePayment:
      description: a payment piloted by Stripe platform.
      type: object
      properties:
        id:
          description: "the Stripe payment id given by viandeendirect.eu"
          x-field-extra-annotation: '@jakarta.persistence.Id'
          type: integer
        checkoutSessionId:
          description: the Stripe Checkout Session id of the payment
          type: string
        paymentUrl:
          type: string
          description: the url where to redirect the customer to process the payment on Stripe platform.
          x-field-extra-annotation: '@jakarta.persistence.Transient'
    Producer:
      description: a producer
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
          description: ""
          x-field-extra-annotation: '@jakarta.persistence.OneToOne'
        id:
          type: integer
          x-field-extra-annotation: '@jakarta.persistence.Id'
        status:
          $ref: '#/components/schemas/ProducerStatus'
        salesCredits:
          description: "the number of sales available"
          type: integer
        productions:
          type: array
          items:
            $ref: '#/components/schemas/Production'
          x-field-extra-annotation: '@jakarta.persistence.OneToMany(mappedBy = "producer")'
        sales:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
          x-field-extra-annotation: '@jakarta.persistence.OneToMany(mappedBy = "seller")'
        stripeAccount:
          $ref: '#/components/schemas/StripeAccount'
      x-class-extra-annotation: '@jakarta.persistence.Entity @jakarta.persistence.Table(name = "producers")'
...
    ProducerStatus:
      type: string
      enum: [ON_BOARDING, ACTIVE]
    OrderStatus:
      type: string
      enum: [ITEMS_SELECTED, PAYMENT_COMPLETED, PAYEMENT_VALIDATED, PAYMENT_FAILED, DELIVERED]